pyproject.toml
[tool.poetry]
name = "waves"
version = "1.3"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pyserial = "^3.5"
comtypes = "^1.4.10"
pycaw = "^20240210"
pyyaml = "^6.0.1"
pydantic = "^2.11.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
pyinstaller = "^6.12.0"

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
